cluster_outage_simulation
