+++ internal/service/clusteroutagesimulation/data_source_cluster_outage_simulation_test.go
@@ -1,22 +1,18 @@
-package mongodbatlas_test
+package clusteroutagesimulation_test
 
 import (
-	"context"
 	"fmt"
 	"os"
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-testing/helper/acctest"
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
-	"github.com/hashicorp/terraform-plugin-testing/terraform"
-	"github.com/mongodb/terraform-provider-mongodbatlas/internal/common/conversion"
-	"github.com/mongodb/terraform-provider-mongodbatlas/internal/config"
 	"github.com/mongodb/terraform-provider-mongodbatlas/internal/testutil/acc"
 )
 
-func TestAccOutageSimulationCluster_SingleRegion_basic(t *testing.T) {
+func TestAccOutageSimulationClusterDS_SingleRegion_basic(t *testing.T) {
 	var (
-		dataSourceName = "mongodbatlas_cluster_outage_simulation.test_outage"
+		dataSourceName = "data.mongodbatlas_cluster_outage_simulation.test"
 		orgID          = os.Getenv("MONGODB_ATLAS_ORG_ID")
 		projectName    = acctest.RandomWithPrefix("test-acc-project")
 		clusterName    = acctest.RandomWithPrefix("test-acc-cluster")
@@ -28,7 +24,7 @@ func TestAccOutageSimulationCluster_SingleRegion_basic(t *testing.T) {
 		CheckDestroy:             testAccCheckMongoDBAtlasClusterOutageSimulationDestroy,
 		Steps: []resource.TestStep{
 			{
-				Config: testAccDataSourceMongoDBAtlasClusterOutageSimulationConfigSingleRegion(projectName, orgID, clusterName),
+				Config: testAccDataSourceMongoDBAtlasClusterOutageSimulationConfigDSSingleRegion(projectName, orgID, clusterName),
 				Check: resource.ComposeTestCheckFunc(
 					resource.TestCheckResourceAttr(dataSourceName, "cluster_name", clusterName),
 					resource.TestCheckResourceAttrSet(dataSourceName, "project_id"),
@@ -42,9 +38,9 @@ func TestAccOutageSimulationCluster_SingleRegion_basic(t *testing.T) {
 	})
 }
 
-func TestAccOutageSimulationCluster_MultiRegion_basic(t *testing.T) {
+func TestAccOutageSimulationClusterDS_MultiRegion_basic(t *testing.T) {
 	var (
-		dataSourceName = "mongodbatlas_cluster_outage_simulation.test_outage"
+		dataSourceName = "data.mongodbatlas_cluster_outage_simulation.test"
 		orgID          = os.Getenv("MONGODB_ATLAS_ORG_ID")
 		projectName    = acctest.RandomWithPrefix("test-acc-project")
 		clusterName    = acctest.RandomWithPrefix("test-acc-cluster")
@@ -56,7 +52,7 @@ func TestAccOutageSimulationCluster_MultiRegion_basic(t *testing.T) {
 		CheckDestroy:             testAccCheckMongoDBAtlasClusterOutageSimulationDestroy,
 		Steps: []resource.TestStep{
 			{
-				Config: testAccDataSourceMongoDBAtlasClusterOutageSimulationConfigMultiRegion(projectName, orgID, clusterName),
+				Config: testAccDataSourceMongoDBAtlasClusterOutageSimulationConfigDSMultiRegion(projectName, orgID, clusterName),
 				Check: resource.ComposeTestCheckFunc(
 					resource.TestCheckResourceAttr(dataSourceName, "cluster_name", clusterName),
 					resource.TestCheckResourceAttrSet(dataSourceName, "project_id"),
@@ -70,7 +66,7 @@ func TestAccOutageSimulationCluster_MultiRegion_basic(t *testing.T) {
 	})
 }
 
-func testAccDataSourceMongoDBAtlasClusterOutageSimulationConfigSingleRegion(projectName, orgID, clusterName string) string {
+func testAccDataSourceMongoDBAtlasClusterOutageSimulationConfigDSSingleRegion(projectName, orgID, clusterName string) string {
 	return fmt.Sprintf(`
 	resource "mongodbatlas_project" "outage_project" {
 		name   = "%s"
@@ -94,10 +90,18 @@ func testAccDataSourceMongoDBAtlasClusterOutageSimulationConfigSingleRegion(proj
 		  region_name    = "US_EAST_1"
 		}
 	}
+
+	data "mongodbatlas_cluster_outage_simulation" "test" {
+		project_id = mongodbatlas_project.outage_project.id
+		cluster_name = mongodbatlas_cluster.atlas_cluster.name
+		depends_on = [
+    		mongodbatlas_cluster_outage_simulation.test_outage,
+  		]
+	}
 	`, projectName, orgID, clusterName)
 }
 
-func testAccDataSourceMongoDBAtlasClusterOutageSimulationConfigMultiRegion(projectName, orgID, clusterName string) string {
+func testAccDataSourceMongoDBAtlasClusterOutageSimulationConfigDSMultiRegion(projectName, orgID, clusterName string) string {
 	return fmt.Sprintf(`
 	resource "mongodbatlas_project" "outage_project" {
 		name   = "%s"
@@ -147,23 +151,13 @@ func testAccDataSourceMongoDBAtlasClusterOutageSimulationConfigMultiRegion(proje
 			   region_name    = "US_EAST_2"
 		}
 	}
-	`, projectName, orgID, clusterName)
-}
 
-func testAccCheckMongoDBAtlasClusterOutageSimulationDestroy(s *terraform.State) error {
-	conn := acc.TestAccProviderSdkV2.Meta().(*config.MongoDBClient).Atlas
-
-	for _, rs := range s.RootModule().Resources {
-		if rs.Type != "mongodbatlas_cluster_outage_simulation" {
-			continue
-		}
-
-		ids := conversion.DecodeStateID(rs.Primary.ID)
-		_, _, err := conn.ClusterOutageSimulation.GetOutageSimulation(context.Background(), ids["project_id"], ids["cluster_name"])
-		if err == nil {
-			return fmt.Errorf("cluster outage simulation for project (%s) and cluster (%s) still exists", ids["project_id"], ids["cluster_name"])
-		}
+	data "mongodbatlas_cluster_outage_simulation" "test" {
+		project_id = mongodbatlas_project.outage_project.id
+		cluster_name = mongodbatlas_cluster.atlas_cluster.name
+		depends_on = [
+    		mongodbatlas_cluster_outage_simulation.test_outage,
+  		]
 	}
-
-	return nil
+	`, projectName, orgID, clusterName)
 }
